Wait For Instance: make it backward compatible & migrate tests (#21)
With the introduction of the Namespace field for GuestAttribute the
configurations without the field namely those with namespace and key
composed in the KeyName field (i.e namespace/key) would be broken.

This patch changes the implementation to not assume Namespace is
always present and introduce tests using the newly added Namespace
field.