Add WaitForAvailableQuotas workflow step and documentation (#30)
This step will take a list of desired quantities of given quotas in given regions, and return when all requested quota is available

As part of this change, add GetRegions capability to the daisy compute client and the test client.

Add an invalidInputError error type to compare error types in unit tests for this step.